apply naming and imports changes from CR
