+++ apps/site/src/main.tsx
@@ -1,25 +1,15 @@
 import "./i18n";
 
-import { Box, Button, Checkbox, Slider } from "@mui/material";
-import { ThemeSelectorButton, ThemeSelectorProvider } from "@patronage-web/features-feedback";
+import { Box } from "@mui/material";
+import { ThemeProvider } from "@patronage-web/features-feedback";
 import { StrictMode } from "react";
 import * as ReactDOM from "react-dom";
 
 ReactDOM.render(
   <StrictMode>
-    <ThemeSelectorProvider>
-      <ThemeSelectorButton />
+    <ThemeProvider>
       <Box>Patronage App</Box>
-      <Box>
-        Theme Selector Demo
-        <Box>
-          <Checkbox />
-          <Button>Test</Button>
-          <Button variant='contained'>Test</Button>
-          <Slider />
-        </Box>
-      </Box>
-    </ThemeSelectorProvider>
+    </ThemeProvider>
   </StrictMode>,
   document.getElementById("root")
 );
