+++ libs/shared/src/components/index.ts
@@ -1,1 +1,1 @@
-export * from "./Themes";
+export * from ".";
