+++ libs/shared/src/utils/get-theme-config.tsx
@@ -1,14 +1,14 @@
 import { grey } from "@mui/material/colors";
 import { createTheme } from "@mui/material/styles";
 
-export const darkTheme = createTheme({
+export const dark = createTheme({
   palette: {
     mode: "dark",
     primary: { main: grey[100] }
   }
 });
 
-export const lightTheme = createTheme({
+export const light = createTheme({
   palette: {
     mode: "light",
     primary: { main: grey[800] }
