+++ libs/shared/src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./get-theme-config";
 export * from "./load-i18n";
